requirements.txt
asgiref==3.3.4
attrs==21.2.0
backports.entry-points-selectable==1.1.0
distlib==0.3.2
dj-static==0.0.6
Django==3.2
django-ckeditor==6.1.0
django-crispy-forms==1.11.2
django-debug-toolbar==3.2.1
django-filter==2.4.0
django-js-asset==1.2.2
djangorestframework==3.12.4
filelock==3.0.12
gunicorn==20.1.0
ipython-genutils==0.2.0
platformdirs==2.0.2
pyrsistent==0.18.0
pytube==10.9.3
pytube3==9.6.4
pytz==2021.1
pywin32==301
pyzmq==22.2.1
six==1.16.0
sqlparse==0.4.1
static3==0.7.0
tornado==6.1
traitlets==5.0.5
typing-extensions==3.10.0.0
virtualenv==20.6.0
wcwidth==0.2.5
